feat: update the design of faq accordion
